+++ schedule/src/app/schedule/add/add.component.html
@@ -51,7 +51,6 @@
         </small>
       </div>
     </div>
-
   </div>
 
   <!--选择时间-->
