+++ schedule/src/app/schedule/edit/edit-clazzes-and-teachers/edit-clazzes-and-teachers.component.html
@@ -1,0 +1,32 @@
+<div class="form-group row mb-0" *ngFor="let clazz of this.schedule.clazzes; index as x">
+  <label class="col-sm-2 text-right col-form-label" style="width: 162px"><code>*</code>班级{{ x + 1 }}</label>
+  <div class="col-sm">
+    <input type="text" class="disabled mt-1" disabled [value]="clazz.name">
+    <button type="button" class="btn btn-danger btn-sm ml-1">删除</button>
+  </div>
+</div>
+<div class="form-group row">
+  <label class="col-sm-2 text-right col-form-label"><code>*</code>上课班级<br>(可多选)</label>
+  <div class="col-sm">
+    <select class="select picker form-control col-sm-3" style="width: 200px" multiple>
+      <option [value]="null">请选择班级</option>
+      <option *ngFor="let clazz of this.thirdFilerClazzes">{{ clazz.name }}</option>
+    </select>
+  </div>
+</div>
+<div class="form-group row">
+  <label class="col-sm-2 text-right col-form-label" style="width: 162px"><code>*</code>教师</label>
+  <div class="col-sm-2">
+    <input type="text" class="disabled mt-1" disabled [value]="2">
+  </div>
+  <div class="col-sm-2">
+    <input type="text" class="disabled mt-1" disabled [value]="3">
+  </div>
+</div>
+<div class="form-group row">
+  <label class="col-sm-2 text-right col-form-label" style="width: 162px"></label>
+  <div class="col-sm">
+    <a type="button" class="btn btn-primary mt-1" routerLink="../">返回</a>
+    <button type="button" class="btn btn-primary mt-1"><i class="fas fa-save"></i>保存修改</button>
+  </div>
+</div>
